python/code-snippetes/kafka-snippets/requirements.txt
confluent-kafka
kafka-python-ng
pandas
